+++ package.json
@@ -26,6 +26,7 @@
     "express": "^4.13.3",
     "flux": "^2.1.1",
     "history": "^1.13.1",
+    "kerberos": "0.0.17",
     "keymirror": "^0.1.1",
     "localStorage": "^1.0.3",
     "material-ui": "^0.13.4",
@@ -40,9 +41,7 @@
     "react-dom": "^0.14.3",
     "react-router": "^1.0.0",
     "react-tap-event-plugin": "^0.2.1",
-    "redis": "^2.4.2",
-    "webpack": "^1.12.9",
-    "webpack-stream": "^2.2.0"
+    "redis": "^2.4.2"
   },
   "devDependencies": {
     "babel-core": "^5.8.24",
@@ -52,6 +51,7 @@
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^2.0.4",
     "gulp-sourcemaps": "^1.6.0",
+    "source-map-loader": "^0.1.5",
     "webpack": "^1.12.2",
     "webpack-stream": "^2.1.1"
   }
