+++ src/components/Home.jsx
@@ -1,12 +1,10 @@
 import React from 'react'
-import { findDOMNode } from 'react-dom'
 import employeeActions from '../actions/employeeActionCreators'
 
 class Home extends React.Component {
   constructor(props) {
     super(props)
     this._login = this._login.bind(this)
-    // this.props.history.replaceState(null, '/home')
   }
 
   render() {
@@ -21,8 +19,12 @@ class Home extends React.Component {
     )
   }
 
+  componentWillMount() {
+    console.log('working')
+  }
+
   _login() {
-    console.log(this.refs.email)
+    console.log(this.refs)
     // if(this.refs.email && this.refs.password) {
     //   employeeActions.login({
     //     email: this.refs.email,
