+++ src/components/Main.jsx
@@ -4,7 +4,6 @@ import { Link } from 'react-router'
 class Main extends React.Component {
   constructor(props) {
     super(props)
-    // this.props.history.replaceState(null, '/')
   }
 
   render() {
